feat: move gas manager funcs to aa infra
