+++ docs/docs.yml
@@ -1161,6 +1161,8 @@ navigation:
                 contents:
                   - page: alchemyEstimateFeesPerGas
                     path: wallets/pages/reference/alchemy/aa-infra/functions/alchemyEstimateFeesPerGas.mdx
+                  - page: alchemyGasManagerHooks
+                    path: wallets/pages/reference/alchemy/aa-infra/functions/alchemyGasManagerHooks.mdx
                   - page: createAlchemyBundlerClient
                     path: wallets/pages/reference/alchemy/aa-infra/functions/createAlchemyBundlerClient.mdx
                   - page: gasManagerActions
@@ -1213,8 +1215,6 @@ navigation:
             contents:
               - section: Functions
                 contents:
-                  - page: alchemyGasManagerHooks
-                    path: wallets/pages/reference/alchemy/smart-accounts/functions/alchemyGasManagerHooks.mdx
                   - page: buildDeferredActionDigest
                     path: wallets/pages/reference/alchemy/smart-accounts/functions/buildDeferredActionDigest.mdx
                   - page: buildFullNonceKey
