+++ docs/pages/reference/alchemy/aa-infra/functions/alchemyGasManagerHooks.mdx
@@ -2,7 +2,7 @@
 # This file is autogenerated
 title: alchemyGasManagerHooks
 description: Overview of the alchemyGasManagerHooks method
-slug: wallets/reference/alchemy/smart-accounts/functions/alchemyGasManagerHooks
+slug: wallets/reference/alchemy/aa-infra/functions/alchemyGasManagerHooks
 ---
 
 Adapts Alchemy's Gas Manager to viem's paymaster interface using the optimized
@@ -20,7 +20,7 @@ for permit support.
 ## Import
 
 ```ts
-import { alchemyGasManagerHooks } from "@alchemy/smart-accounts";
+import { alchemyGasManagerHooks } from "@alchemy/aa-infra";
 ```
 
 ## Usage
