+++ packages/smart-accounts/package.json
@@ -44,7 +44,6 @@
     "typescript-template": "*"
   },
   "dependencies": {
-    "@alchemy/aa-infra": "*",
     "@alchemy/common": "*",
     "ox": "^0.8.5",
     "viem": "2.29.2"
